magit-show-commit: when in a log buffer, use log's file restriction
All calls to magit-show-commit take the file restriction from
magit-diff-arguments.  But when a call is made from a log buffer, the
user is viewing the revision in the context of the log, so the log's
file restriction is more relevant than the diff's [*].

Adjust magit-show-commit calls to respect the log's file restriction
if the current buffer is a log buffer.  Leave calls to
magit-show-commit in git-rebase.el and magit-blame.el unchanged
because these aren't designed to be executed from a log buffer.

Closes #3056

[*] This patch assumes that all users would prefer the log's file
    restriction over the diff's in this situation, but of course we
    can add an option to revert to the previous behavior if this turns
    out not to be the case.